+++ components/tbt-dropdown.js
@@ -1,6 +1,6 @@
 /**
  * @component tbt-dropdown
- * @version 1.46.0
+ * @version 1.46.1
  * @author Wichit Wongta
  *
  * Styled select dropdown with label, placeholder, and validation.